requirements.txt
label-studio-sdk>=0.0.23
numpy>=1.21.0
Pillow>=8.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
requests>=2.0.0 